backend/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv
langchain-pinecone
langchain_community
langchain_google_genai
langchain_experimental
requests==2.31.0
python-docx
python-pptx
